创建入库/出库单返回自增id

+++ jshERP-boot/src/main/java/com/jsh/erp/service/stockIn/StockInBillService.java
@@ -13,7 +13,7 @@ public interface StockInBillService {
 
     List<StockInDetailVo> queryStockInDetail(StockInBillVo stockInBillVo);
 
-    int add(StockInBillVo stockInBillVo);
+    long add(StockInBillVo stockInBillVo);
 
     boolean delete(long id);
 }
